Minor: Apply alerts in Authentication
